third_party/blink/web_tests/fast/css/computed-offset-with-zoom-expected.txt

PASS window.getComputedStyle(fixedElement).top is '100px'
PASS window.getComputedStyle(fixedElement).right is '100px'
PASS window.getComputedStyle(fixedElement).bottom is '100px'
PASS window.getComputedStyle(fixedElement).left is '100px'
PASS window.getComputedStyle(percentageElement).top is '50px'
PASS window.getComputedStyle(percentageElement).right is '100px'
PASS window.getComputedStyle(percentageElement).bottom is '50px'
PASS window.getComputedStyle(percentageElement).left is '100px'
PASS window.getComputedStyle(viewportElement).top is '60px'
PASS window.getComputedStyle(viewportElement).right is '80px'
PASS window.getComputedStyle(viewportElement).bottom is '60px'
PASS window.getComputedStyle(viewportElement).left is '80px'
PASS window.getComputedStyle(fixedElement).top is '100px'
PASS window.getComputedStyle(fixedElement).right is '100px'
PASS window.getComputedStyle(fixedElement).bottom is '100px'
PASS window.getComputedStyle(fixedElement).left is '100px'
PASS window.getComputedStyle(percentageElement).top is '50px'
PASS window.getComputedStyle(percentageElement).right is '100px'
PASS window.getComputedStyle(percentageElement).bottom is '50px'
PASS window.getComputedStyle(percentageElement).left is '100px'
PASS window.getComputedStyle(viewportElement).top is '60px'
PASS window.getComputedStyle(viewportElement).right is '80px'
PASS window.getComputedStyle(viewportElement).bottom is '60px'
PASS window.getComputedStyle(viewportElement).left is '80px'
PASS window.getComputedStyle(fixedElement).top is '100px'
PASS window.getComputedStyle(fixedElement).right is '100px'
PASS window.getComputedStyle(fixedElement).bottom is '100px'
PASS window.getComputedStyle(fixedElement).left is '100px'
PASS window.getComputedStyle(percentageElement).top is '50px'
PASS window.getComputedStyle(percentageElement).right is '100px'
PASS window.getComputedStyle(percentageElement).bottom is '50px'
PASS window.getComputedStyle(percentageElement).left is '100px'
PASS window.getComputedStyle(viewportElement).top is '60px'
PASS window.getComputedStyle(viewportElement).right is '80px'
PASS window.getComputedStyle(viewportElement).bottom is '60px'
PASS window.getComputedStyle(viewportElement).left is '80px'
PASS successfullyParsed is true

TEST COMPLETE


Reported computed offset should not change with zoom level.
